Separated Library and added Interruptions
Added partial Interrupt support to the library.
Not all Interruptions present in the interrupt vector were implemented because there are a lot, but the interruptions were simplified.
Also separated the library from the project files so that every project uses the same library. This library has to be compiled every time changes are made