Updated the examples to include low-power
Updated the blueprint so that the function "alwaysSetup" disables every module by default (using the PMDx registers), having the setup functions on the library enable the peripheral they're setting up.
Updated the LED_Blink and the ADC examples so that they include this feature.

The LED_Blink example also enters idle mode, being interrupted to blink the LED. Sleep mode is better because it disables everything but that is the problem, it even disables the clock powering the Timer1 which was the only thing waking it up (via its interrupt). Sleep mode can be used only if there is an external clock powering Timer1.