Code working without integrating the AXL
Only thing left for 1st delivery is to integrate the accelerometer.
Created a project on delivery 1 for the master board, with multiple commands to ask the SPI slave for different values with varying results... Probably because of SPI timings... Not something to investigate for 1st delivery. Next step is to integrate the accelerometer